Keep next/headers out of the client bundle
The hook that records the sort imported the shared module, which pulled
cookies() into a client component and broke the build. The cookie read
moves to its own server-only module; the constants and parsers stay
shared.
