+++ src/lib/list-sort-preference.ts
@@ -1,5 +1,3 @@
-import { cookies } from 'next/headers'
-
 /**
  * Remembered sort order for the list pages.
  *
@@ -61,26 +59,4 @@ export function parseStoredSort(stored: string | undefined): ListSort | null {
   const [sortBy, sortOrder] = stored.split(':')
   if (!sortBy) return null
   return { sortBy, sortOrder: sortOrder === 'asc' ? 'asc' : 'desc' }
-}
-
-/**
- * The sort a list page should render with: an explicit `?sortBy=` in the URL
- * always wins, then the remembered choice, then the page's own default.
- *
- * The URL taking precedence matters — a shared link, a back button and a
- * column click all put the order in the URL, and none of them should be
- * overridden by what this browser happened to pick last.
- */
-export async function resolveListSort(
-  list: ListKey,
-  params: { sortBy?: string; sortOrder?: string },
-  fallback: ListSort = { sortBy: undefined, sortOrder: 'desc' }
-): Promise<ListSort> {
-  if (params.sortBy) {
-    return { sortBy: params.sortBy, sortOrder: params.sortOrder === 'asc' ? 'asc' : 'desc' }
-  }
-
-  const store = await cookies()
-  const remembered = parseStoredSort(parseListSortCookie(store.get(LIST_SORT_COOKIE)?.value)[list])
-  return remembered ?? fallback
 }
